+++ lib/structure.rb
@@ -60,6 +60,8 @@ module Structure
 
       # parse accepts JSON-ish hashes + kwargs override
       klass.define_singleton_method(:parse) do |data = {}, **kwargs|
+        return data if data.is_a?(self)
+
         string_kwargs = kwargs.transform_keys(&:to_s)
         data = data.merge(string_kwargs)